refactor: move definitions out of headers.
