+++ sat.h
@@ -1,10 +1,7 @@
-#ifndef NETPBM_INPUT_SAT_H_
-#define NETPBM_INPUT_SAT_H_
+#ifndef NETPBM__SAT_H_
+#define NETPBM__SAT_H_
 
-#include <malloc.h>
-#include <stddef.h>
 #include <stdint.h>
-#include <stdio.h>
 
 #include "types/pgm.h"
 #include "types/sat.h"
@@ -17,24 +14,7 @@
  * @return          A pointer to the SummedAreaTable, or NULL if an error
  * occurred.
  */
-SummedAreaTable *AllocateSat(uint32_t width, uint32_t height) {
-  // Allocate memory for image data
-  SummedAreaTable *sat = (SummedAreaTable *)malloc(sizeof(SummedAreaTable));
-  if (!sat) {
-    fprintf(stderr, "Error: out of memory\n");
-    return NULL;
-  }
-  sat->width_  = width;
-  sat->height_ = height;
-  sat->data_   = (uint64_t *)calloc(width * height, sizeof(uint64_t));
-  if (!sat->data_) {
-    fprintf(stderr, "Error: out of memory\n");
-    free(sat);
-    return NULL;
-  }
-
-  return sat;
-}
+extern SummedAreaTable *AllocateSat(uint32_t width, uint32_t height);
 
 /**
  * Compute the summed area table of a PGM image.
@@ -42,39 +22,7 @@ SummedAreaTable *AllocateSat(uint32_t width, uint32_t height) {
  * @param pgm   The PGM image.
  * @return      The summed area table, or NULL if an error occurred.
  */
-SummedAreaTable *PgmToSat(PgmImage *pgm) {
-  uint32_t width  = pgm->width_;
-  uint32_t height = pgm->height_;
-
-  // Allocate memory for summed area table
-  SummedAreaTable *sat = AllocateSat(width, height);
-  if (!sat) {
-    return NULL;
-  }
-
-#pragma omp parallel for default(none) shared(sat, pgm, width)
-  // Copy first row
-  for (uint32_t x = 0; x < width; x++) {
-    sat->data_[x] = (uint64_t)pgm->data_[x];
-  }
-
-  // Column-wise sum
-  for (uint32_t y = 1; y < height; y++) {
-#pragma omp parallel for default(none) shared(sat, pgm, width, y)
-    for (uint32_t x = 0; x < width; x++) {
-      sat->data_[y * width + x] =
-          (uint64_t)pgm->data_[y * width + x] + sat->data_[(y - 1) * width + x];
-    }
-  }
-#pragma omp parallel for default(none) shared(sat, width, height)
-  //  Row-wise sum
-  for (uint32_t y = 0; y < height; y++) {
-    for (uint32_t x = 1; x < width; x++) {
-      sat->data_[y * width + x] += sat->data_[y * width + x - 1];
-    }
-  }
-  return sat;
-}
+extern SummedAreaTable *PgmToSat(const PgmImage *pgm);
 
 /**
  * Query the summed area table.
@@ -87,23 +35,14 @@ SummedAreaTable *PgmToSat(PgmImage *pgm) {
  * @return      The sum of the pixels in the rectangle defined by the given
  * coordinates.
  */
-uint64_t SatQuery(SummedAreaTable *sat, uint32_t tlx, uint32_t tly,
-                  uint32_t brx, uint32_t bry) {
-  uint64_t res = sat->data_[bry * sat->width_ + brx];
-  if (tly > 0) res -= sat->data_[(tly - 1) * sat->width_ + brx];
-  if (tlx > 0) res -= sat->data_[bry * sat->width_ + tlx - 1];
-  if (tlx > 0 && tly > 0) res += sat->data_[(tly - 1) * sat->width_ + tlx - 1];
-  return res;
-}
+extern uint64_t SatQuery(const SummedAreaTable *sat, uint32_t tlx, uint32_t tly,
+                         uint32_t brx, uint32_t bry);
 
 /**
  * Free memory for a summed area table.
  *
  * @param sat  The summed area table to free.
  */
-void FreeSat(SummedAreaTable *sat) {
-  free(sat->data_);
-  free(sat);
-}
+extern void FreeSat(SummedAreaTable *sat);
 
-#endif// NETPBM_INPUT_SAT_H_
+#endif// NETPBM__SAT_H_
